BaseResult拆分

+++ goperation/manager/models.py
@@ -259,7 +259,7 @@ class AgentReportLog(PluginTableBase):
     buffers = sa.Column(INTEGER(unsigned=True), nullable=False)
     free = sa.Column(INTEGER(unsigned=True), nullable=False)
     # partion left size MB
-    left = sa.Column(INTEGER(unsigned=True), nullable=False)
+    left = sa.Column(BIGINT(unsigned=True), nullable=False)
     # network  count
     # psutil.net_connections()  count(*)
     syn = sa.Column(INTEGER(unsigned=True), nullable=False)
